+++ src/lib/components/TimeLine.svelte
@@ -1,0 +1,30 @@
+<script lang="ts">
+	import Note from '$lib/components/Note/Note.svelte';
+	import { subscribe } from '$lib/utils/nostr';
+	import type { Event, Filter } from 'nostr-tools';
+	import { onMount } from 'svelte';
+	import { writable } from 'svelte/store';
+
+	export let authors: string[];
+	export let filter: Omit<Filter, 'kinds' | 'authors'> = {};
+
+	const notes = writable<Event[]>([]);
+
+	onMount(() => {
+		const sub = subscribe({
+			...filter,
+			kinds: [1],
+			authors
+		});
+
+		sub.on('event', (event: Event) => {
+			notes.update((updater) =>
+				[...updater, event].sort((a, b) => b.created_at - a.created_at)
+			);
+		});
+	});
+</script>
+
+{#each $notes as note (note.id)}
+	<Note {note} />
+{/each}
